% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk1-models.R
\docType{data}
\name{pk1-models}
\alias{pk1-models}
\alias{pk1.fixed.ode.noise}
\alias{pk1.fixed.sde.noise}
\alias{pk1.fixed.sde.pure}
\alias{pk1.mixed.ode.noise}
\alias{pk1.mixed.sde.noise}
\alias{pk1.mixed.sde.pure}
\title{Variations on the One-Compartment Pharmacokinetic (PK1) Model.}
\format{An object of class \code{stanmodel} of length 1.}
\usage{
sampling(object = pk1.mixed.sde.noise, ...)

sampling(object = pk1.mixed.sde.pure, ...)

sampling(object = pk1.mixed.ode.noise, ...)

sampling(object = pk1.fixed.ode.noise, ...)

sampling(object = pk1.fixed.sde.pure, ...)

sampling(object = pk1.fixed.sde.noise, ...)
}
\description{
Stan implementations of the PK1 model, with ODE or SDE formulation, fixed or mixed effects, and with or without additional measurement error.
}
\details{
Each of these objects are of class \code{stanmodel}, such that parameter inference via MCMC can be done with the \code{rstan} package, e.g., with the call

\code{sampling(object = stanmodel, iter = 1e4, chains = 1)}
}
\keyword{datasets}

