% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk1-functions.R
\name{pk1.Cmax.sim}
\alias{pk1.Cmax.sim}
\title{Simulation of maximal concentration for the PK1 model.}
\usage{
pk1.Cmax.sim(tObs, X0, Dose, Ke, Ka, Cl, sigmaP)
}
\arguments{
\item{tObs}{vector of observation times.}

\item{X0}{Initial concentration (at time tObs[1]).}

\item{Dose}{Administered drug dosage.}

\item{Ke}{Elimination rate.}

\item{Ka}{Absorption rate.}

\item{Cl}{Clearance level.}

\item{sigmaP}{Brownian force scaling.  \code{sigmaP = 0} solve the ODE instead.}
}
\value{
A matrix with columns \code{Cmax} and \code{tCmax}.
}
\description{
Simulation of maximal concentration for the PK1 model.
}
\details{
Generates a concentration curve at specified timepoints, and records the maximal concentration value along with the time when it happens.
}

