% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk1-functions.R
\name{pk1.init}
\alias{pk1.init}
\title{Initialize the stan sampler}
\usage{
pk1.init(descr = NULL, effect = c("Fixed", "Mixed", "Indep"),
  DE = c("SDE", "ODE"), meas = c("Pure", "Noise"), Cl, Ka, Ke, sigmaP,
  sigmaM, muCl, muKa, muKe, sigCl, sigKa, sigKe, Xt)
}
\description{
Initialize the stan sampler
}
