% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk1-functions.R
\name{pk1.loglik}
\alias{pk1.loglik}
\title{Loglikelihood of the PK1 Model}
\usage{
pk1.loglik(Cl, Ka, Ke, sigmaP, sigmaM, yObs, tObs, Dose)
}
\arguments{
\item{Cl}{Clearance level.}

\item{Ka}{Absorption rate.}

\item{Ke}{Elimination rate.}

\item{sigmaP}{Brownian force scaling.  \code{sigmaP = 0} solve the ODE instead.}

\item{sigmaM}{Standard deviation of measurment error.}

\item{yObs}{Vector of measurements.}

\item{tObs}{Vector of observation times.}

\item{Dose}{Administered drug dosage.}
}
\value{
If \code{yObs} is not missing, loglikelihood.  Otherwise, list with elements \code{mu} and \code{V} of \code{yObs}.
}
\description{
Calculates the loglikelihood for different formulations of the fixed-effects PK1 model, i.e., as an SDE or ODE, and with or without measurement error.
}
