% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk1-functions.R
\name{pk1.ode}
\alias{pk1.ode}
\title{Solution of the PK1 ODE model.}
\usage{
pk1.ode(tObs, Dose, Cl, Ka, Ke)
}
\arguments{
\item{tObs}{Vector of observation times.}

\item{Dose}{Administered drug dosage.}

\item{Cl}{Clearance level.}

\item{Ka}{Absorption rate.}

\item{Ke}{Elimination rate.}
}
\value{
Solution vector at times \code{tObs}.
}
\description{
Solution of the PK1 ODE model.
}
\details{
The initial concentration (at time \code{t = 0}) is assumed to be \code{X0 = 0}.
}
