% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk1-functions.R
\name{pk1.sim}
\alias{pk1.sim}
\title{Simulation of the PK1 SDE model with unequal sampling times.}
\usage{
pk1.sim(tObs, X0, Dose, Ke, Ka, Cl, sigmaP)
}
\arguments{
\item{tObs}{Vector of observation times.}

\item{X0}{Initial concentration (at time \code{tObs[1]}).}

\item{Dose}{Administered drug dosage.}

\item{Ke}{Elimination rate.}

\item{Ka}{Absorption rate.}

\item{Cl}{Clearance level.}

\item{sigmaP}{Brownian force scaling.  \code{sigmaP = 0} reduces to the corresponding ODE.}
}
\value{
Concentration matrix of size \code{nReps x nObs}, where \code{nObs = length(tObs)} and \code{nReps} is determined by the length of the other inputs.
}
\description{
Simulation of the PK1 SDE model with unequal sampling times.
}
\details{
Each of \code{X0}, \code{Dose}, \code{Ke}, \code{Ka}, \code{Cl}, and \code{sigmaP} can be scalars or vectors.  In case of the latter, a concentration curve is returned for each value.
}
