% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gAR1.R
\name{gAR1.MV}
\alias{gAR1.MV}
\title{Mean and Variance of a Generalized AR(1) Process}
\usage{
gAR1.MV(aa, bb, cc, x0, debug = FALSE)
}
\value{
A list with elements \code{mu} and \code{V} if \code{x0} is given.  Otherwise, a list with elements \code{A}, \code{b} and \code{V}, where \code{mu = A x0 + b}.
}
\description{
Mean and Variance of a Generalized AR(1) Process
}
\details{
The GAR(1) process is given by
\deqn{x[n] = a[n] * x[n-1] + b[n] + c[n] * e[n], \qquad e[n] ~iid N(0,1)}.
}

