% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk1-functions.R
\name{pk1.ode}
\alias{pk1.ode}
\title{PK1 ODE Solution}
\usage{
pk1.ode(tObs, Dose, Cl, Ka, Ke)
}
\arguments{
\item{tObs}{vector of observation times.}

\item{Dose}{Administered drug dosage.}

\item{Cl}{Clearance level.}

\item{Ka}{Absorption rate.}

\item{Ke}{Elimination rate.}
}
\description{
PK1 ODE Solution
}

